requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
confluent-kafka==2.3.0
Django==4.2.6
djangorestframework==3.14.0
kafka-python==2.0.2
passlib==1.7.4
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
